pr#2 modification
