+++ furiosa_ssd_tf/datasets/pascalvoc_common.py
@@ -65,7 +65,6 @@ def get_split(split_name, dataset_dir, file_pattern, reader,
         ValueError: if `split_name` is not a valid train/test split.
     """
     if split_name not in split_to_sizes:
-        print(split_to_sizes)
         raise ValueError('split name %s was not recognized.' % split_name)
     file_pattern = os.path.join(dataset_dir, file_pattern % split_name)
